Moved a number of structs to their own file (or the right file)
